Add DGrow ()
The problem is that DDump passes *a copy of* the scalar to DDump_XS
or DDump_IO.  The copying process doesn't preserve the allocated space.
The attached patch fixes this, making t/52_DGrow.t pass (apart from
the incorrect test count), but also makes most of t/30_DDump-s.t fail,
because it was relying on a similar side-effect of the copying.