Adjust for new way Test::More preserves vars
Test::More no longer uses 'local' to protect $!, $? and $@. Instead it
stores the values and restores them in another way. This was discussed
on cpan-workers, and it was noticed that it means that magic flags are
not always preserved. It was decided that this was acceptible.

There is also a patch for p5p that has an almost identical fix to a
nearly identical XS test.